Refactor: Move Hyde facade methods to traits
